feat!: use valkey lib instead of redis
The bot, as of recently, is using valkey in production.
While the bot still uses redis-compatible architecture for now,
there is a high chance that, in the future, that something will break between the two.
This shouldn't result in any changes beyond doing a reinstall of requirements.txt.

The docker image will be changed at a later date.